refactor: Move GlobAll to utils
